requirements.txt
cffi==0.9.2
cryptography==0.8.1
Django==1.7
django-allauth==0.19.1
django-ckeditor==4.4.7
django-pipeline==1.4.7
django-taggit==0.12.3
django-watson==1.1.5
enum34==1.0.4
futures==2.2.0
oauthlib==0.7.2
paypalrestsdk==1.9.0
pyasn1==0.1.7
pycparser==2.10
pyOpenSSL==0.14
python-openid==2.2.5
pytz==2014.10
requests==2.6.0
requests-oauthlib==0.4.2
sorl-thumbnail==12.2
